% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{nse.boot}
\alias{nse.boot}
\title{Bootstrap NSE estimator}
\usage{
nse.boot(x, nb, type = c("stationary", "circular"), b = NULL,
  lag.prewhite = 0)
}
\arguments{
\item{x}{A numeric vector or a matrix.}

\item{nb}{The number of bootstrap replications.}

\item{type}{The bootstrap schemes c("stationary","circular").}

\item{b}{the block length. If NULL automatic block length selection.}

\item{lag.prewhite}{Prewhite the serie before analysis (integer or NULL, i.e. automatic selection).}
}
\value{
The variance estimator in the univariate case or the variance-covariance matrix 
estimator in the multivariate case.
}
\description{
Calculate the variance of the mean with a bootstrap variance estimator.
}
\note{
Use the automatic blocksize in \code{\link[np]{b.star}} from th np package which is based 
on Politis and White (2004) and Patton and al (2009). 
Two bootstrap schemes are available; The stationary bootstrap of Politis and Romano  (1994)
and the circular bootstrap of Politis and Romano (1992).
}
\examples{
 
n    = 1000
ar   = 0.9
mean = 1
sd   = 1

set.seed(1234) 
x = as.vector(arima.sim(n = n, list(ar = ar), sd = sd) + mean)
 
set.seed(1234)
nse.boot(x = x, nb = 1000, type = "stationary", b = NULL, lag.prewhite = 0)
nse.boot(x = x, nb = 1000, type = "stationary", b = NULL, lag.prewhite = 1)
nse.boot(x = x, nb = 1000, type = "stationary", b = NULL, lag.prewhite = NULL)

nse.boot(x = x, nb = 1000, type = "stationary", b = 10, lag.prewhite = 0)
nse.boot(x = x, nb = 1000, type = "stationary", b = 10, lag.prewhite = 1)
nse.boot(x = x, nb = 1000, type = "stationary", b = 10, lag.prewhite = NULL)

nse.boot(x = x, nb = 1000, type = "circular", b = NULL, lag.prewhite = 0)
nse.boot(x = x, nb = 1000, type = "circular", b = NULL, lag.prewhite = 1)
nse.boot(x = x, nb = 1000, type = "circular", b = NULL, lag.prewhite = NULL)

nse.boot(x = x, nb = 1000, type = "circular", b = 10, lag.prewhite = 0)
nse.boot(x = x, nb = 1000, type = "circular", b = 10, lag.prewhite = 1)
nse.boot(x = x, nb = 1000, type = "circular", b = 10, lag.prewhite = NULL)
}
\author{
David Ardia and Keven Bluteau
}
\references{
Politis, D.N., Romano, and J.P. (1992).
A circular block-resampling procedure for stationary data.
In \emph{Exploring the limits of bootstrap}, John Wiley & Sons, pp.263--270.

Politis, D.N., Romano, and J.P. (1994). 
The stationary bootstrap.
\emph{Journal of the American Statistical Association} \bold{89}(428), pp.1303--1313. \doi{10.2307/2290993}.

Politis, D.N., White, H. (2004).
Automatic block-length selection for the dependent bootstrap.
\emph{Econometric Reviews} \bold{23}(1), pp.53--70. \doi{10.1081/etc-120028836}.
}

