% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{nse.hiruk}
\alias{nse.hiruk}
\title{Hirukawa NSE estimator}
\usage{
nse.hiruk(x, type = c("bartlett", "parzen"), lag.prewhite = 0)
}
\arguments{
\item{x}{A numeric vector.}

\item{type}{The type of kernel used c("Bartlett","Parzen").}

\item{lag.prewhite}{Prewhite the serie before analysis (integer or NULL, i.e. automatic selection).}
}
\value{
The variance estimator.
}
\description{
Calculate the variance of the mean with the kernel based variance estimator 
by Andrews (1991) using Hirukawa (2010) automatic bandwidth estimator.
}
\details{
This is a wrapper around \code{\link[sandwich]{lrvar}} from the sandwich package 
and use Hirukawa (2010) automatic bandwidth estimator.
}
\examples{
n    = 1000
ar   = 0.9
mean = 1
sd   = 1

set.seed(1234) 
x = as.vector(arima.sim(n = n, list(ar = ar), sd = sd) + mean)
 
nse.hiruk(x = x, type = "bartlett", lag.prewhite = 0)
nse.hiruk(x = x, type = "bartlett", lag.prewhite = 1)
nse.hiruk(x = x, type = "bartlett", lag.prewhite = NULL)

nse.hiruk(x = x, type = "parzen", lag.prewhite = 0)
nse.hiruk(x = x, type = "parzen", lag.prewhite = 1)
nse.hiruk(x = x, type = "parzen", lag.prewhite = NULL)

}
\author{
David Ardia and Keven Bluteau
}
\references{
Hirukawa, M. (2010). 
A two-stage plug-in bandwidth selection and its implementation for covariance estimation.
\emph{Econometric Theory}, \bold{26}(3), pp.710--743. \doi{10.1017/s0266466609990089}.
}

