% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{nse.nw}
\alias{nse.nw}
\title{Newey-West NSE estimator}
\usage{
nse.nw(x, lag.prewhite = 0)
}
\arguments{
\item{x}{A numeric vector or matrix.}

\item{lag.prewhite}{Prewhite the serie before analysis (integer or NULL, i.e. automatic selection).}
}
\value{
The variance estimator in the univariate case or the variance-covariance matrix 
estimator in the multivariate case.
}
\description{
Calculate the variance of the mean with the Newey West (1987, 1994) HAC estimator.
}
\note{
This is a wrapper around \code{\link[sandwich]{lrvar}} from the sandwich package.
}
\examples{
n    = 1000
ar   = 0.9
mean = 1
sd   = 1

set.seed(1234)   
x = as.vector(arima.sim(n = n, list(ar = ar), sd = sd) + mean)
 
nse.nw(x = x, lag.prewhite = 0)
nse.nw(x = x, lag.prewhite = 1)
nse.nw(x = x, lag.prewhite = NULL)
}
\author{
David Ardia and Keven Bluteau
}
\references{
Newey, W.K., West, K.D. (1987).
A simple, positive semi-definite, heteroskedasticity and autocorrelationconsistent covariance matrix. 
\emph{Econometrica} \bold{55}(3), pp.703--708. \doi{10.2307/1913610}.

Newey, W.K., West, K.D. (1994) .
Automatic lag selection in covariance matrix estimation.
\emph{Review of Economic Studies} \bold{61}(4), pp.631--653. \doi{doi: 10.3386/t0144}
}

