% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{nse.spec0}
\alias{nse.spec0}
\title{Spectral density at zero NSE estimator}
\usage{
nse.spec0(x, type = c("ar", "glm", "wosa", "bartlett", "tukey"),
  lag.prewhite = 0)
}
\arguments{
\item{x}{A numeric vector.}

\item{type}{A character string denoting the method to use in estimating the spectral density function.}

\item{lag.prewhite}{Prewhite the serie before analysis (integer or NULL, i.e. automatic selection).}
}
\value{
The variance estimator.
}
\description{
Calculate the variance of the mean with the spectrum at zero estimator.
}
\note{
This is a wrapper around \code{\link[coda]{spectrum0.ar}} from the CODA package 
and \code{\link[mcmcse]{mcse}} from the mcmcse package.
}
\examples{
n    = 1000
ar   = 0.9
mean = 1
sd   = 1
set.seed(1234)   
x = as.vector(arima.sim(n = n, list(ar = ar), sd = sd) + mean)
 
nse.spec0(x = x, type = "ar", lag.prewhite = 0)
nse.spec0(x = x, type = "ar", lag.prewhite = 1)
nse.spec0(x = x, type = "ar", lag.prewhite = NULL)

nse.spec0(x = x, type = "glm", lag.prewhite = 0)
nse.spec0(x = x, type = "glm", lag.prewhite = 1)
nse.spec0(x = x, type = "glm", lag.prewhite = NULL)

nse.spec0(x = x, type = "wosa", lag.prewhite = 0)
nse.spec0(x = x, type = "wosa", lag.prewhite = 1)
nse.spec0(x = x, type = "wosa", lag.prewhite = NULL)

nse.spec0(x = x, type = "bartlett", lag.prewhite = 0)
nse.spec0(x = x, type = "bartlett", lag.prewhite = 1)
nse.spec0(x = x, type = "bartlett", lag.prewhite = NULL)

nse.spec0(x = x, type = "tukey", lag.prewhite = 0)
nse.spec0(x = x, type = "tukey", lag.prewhite = 1)
nse.spec0(x = x, type = "tukey", lag.prewhite = NULL)
 
}
\author{
David Ardia and Keven Bluteau
}
\references{
Flegal, J.M., Hughes, J., Vats D. (2010). 
Batch means and spectral variance estimators in Markov chain Monte Carlo.
\emph{Annals of Statistics} \bold{38}(2), pp.1034--1070. \doi{10.1214/09-aos735}.
}

