% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{nse.andrews}
\alias{nse.andrews}
\title{Andrews NSE estimators.}
\usage{
nse.andrews(x, type = c("bartlett", "parzen", "tukey", "qs", "trunc"),
  lag.prewhite = 0, approx = c("AR(1)", "ARMA(1,1)"))
}
\arguments{
\item{x}{A numeric vector or matrix.}

\item{type}{The type of kernel used c("bartlett","parzen","qs","trunc","tukey").}

\item{lag.prewhite}{Prewhite the serie before analysis (integer or NULL, i.e. automatic selection)}

\item{approx}{Andrews approximation c("AR(1)", "ARMA(1,1)")}
}
\value{
The variance estimator in the univariate case or the variance-covariance matrix estimator in the multivariate case.
}
\description{
Calculate the variance of the mean with the kernel based variance estimator indtroduced by Andrews (1991).
}
\details{
This is a wrapper around \link[sandwich]{lrvar} from the sandwich package and use Andrews (1991) automatic bandwidth estimator.
}
\examples{
n    = 1000
ar   = 0.9
mean = 1
sd   = 1
 
set.seed(1234)   
x = as.vector(arima.sim(n = n, list(ar = ar), sd = sd) + mean)
 
nse.andrews(x = x, type = "bartlett", lag.prewhite = 0)
nse.andrews(x = x, type = "bartlett", lag.prewhite = 1)
nse.andrews(x = x, type = "bartlett", lag.prewhite = NULL)

nse.andrews(x = x, type = "parzen", lag.prewhite = 0)
nse.andrews(x = x, type = "parzen", lag.prewhite = 1)
nse.andrews(x = x, type = "parzen", lag.prewhite = NULL)

nse.andrews(x = x, type = "tukey", lag.prewhite = 0)
nse.andrews(x = x, type = "tukey", lag.prewhite = 1)
nse.andrews(x = x, type = "tukey", lag.prewhite = NULL)

nse.andrews(x = x, type = "qs", lag.prewhite = 0)
nse.andrews(x = x, type = "qs", lag.prewhite = 1)
nse.andrews(x = x, type = "qs", lag.prewhite = NULL)

nse.andrews(x = x, type = "trunc", lag.prewhite = 0)
nse.andrews(x = x, type = "trunc", lag.prewhite = 1)
nse.andrews(x = x, type = "trunc", lag.prewhite = NULL)
}
\references{
Zeileis, Achim. "Econometric computing with HC and HAC covariance matrix estimators." (2004).

Andrews, Donald WK. "Heteroskedasticity and autocorrelation consistent 
covariance matrix estimation." Econometrica: Journal of the Econometric Society 59.03 (1991): 817-858.

Newey, Whitney K., and Kenneth D. West. "A simple, positive semi-definite, 
heteroskedasticity and autocorrelationconsistent covariance matrix.", Econometrica: Journal of the Econometric Society 55.03 (1987) : 703-708.
}

