% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{nse.boot}
\alias{nse.boot}
\title{Bootstrap NSE estimators.}
\usage{
nse.boot(x, nb, type = "stationary")
}
\description{
Calculate the variance of the mean with a bootstrap variance estimator.
}
\details{
Use the automatic blocksize in \link[np]{b.star} from th np package which is based on Politis and White (2004) and Patton and al (2009). 
Two bootstrap schemes are available; The stationary bootstrap of Politis and Romano  (1994)
and the circular bootstrap of Politis and Romano (1992).
}
\examples{
 
n = 1000
ar = c(0.9,0.6)
mean = c(1,5)
sd = c(10,2)
nb = 100
 
Ts1 = as.vector(arima.sim(n = n, list(ar = ar[1]), sd = sd[1]) + mean[1])
Ts2 = as.vector(arima.sim(n = n, list(ar = ar[2]), sd = sd[2]) + mean[2])
Ts = cbind(Ts1,Ts2)
 
nse::nse.boot(x = Ts1, nb =  nb, type = "stationary")
nse::nse.boot(x = Ts, nb =  nb, type = "stationary")
nse::nse.boot(x = Ts1, nb =  nb, type = "circular")
nse::nse.boot(x = Ts, nb =  nb, type = "circular")
 
    @param x       A numeric vector or a matrix.
    @param nb   The number of bootstrap replication.
    @param type    The bootstrap schemes c("stationary","circular").
    @return The variance estimator in the univariate case or the variance-covariance matrix estimator in the multivariate case.
   @references Politis, Dimitris N., and Joseph P. Romano. "A circular block-resampling procedure for stationary data." Exploring the limits of bootstrap (1992): 263-270.
   @references Politis, Dimitris N., and Halbert White. "Automatic block-length selection for the dependent bootstrap." Econometric Reviews 23.1 (2004): 53-70.
   @references Patton, Andrew, Dimitris N. Politis, and Halbert White. "Correction to "Automatic block-length selection for the dependent bootstrap" by D. Politis and H. White." Econometric Reviews 28.4 (2009): 372-375.
   @references Politis, Dimitris N., and Joseph P. Romano. "The stationary bootstrap." Journal of the American Statistical association 89.428 (1994): 1303-1313.
   @references Hayfield, Tristen, and Jeffrey S. Racine. "Nonparametric econometrics: The np package." Journal of statistical software 27.5 (2008): 1-32.
}

