% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{nse.geyer}
\alias{nse.geyer}
\title{Variance of sample mean of functional of reversible Markov chain using methods of Geyer (1992).}
\usage{
nse.geyer(x, type = c("iseq", "bm", "obm", "iseq.bm"), nbatch = 30,
  iseq.type = c("pos", "dec", "con"))
}
\arguments{
\item{x}{A numeric vector or a matrix(only for type "bm").}

\item{type}{The type c("iseq","bm","obm","iseq.bm").}

\item{nbatch}{An optional parameter for the type m and iseq.bm.}

\item{iseq.type}{constraints on function, ("pos") nonnegative, ("dec") nonnegative 
and nonincreasing, and ("con") nonnegative, nonincreasing, and convex. The default value is "pos".}
}
\value{
The variance estimator in the univariate case or the v
ariance-covariance matrix estimator in the multivariate case.
}
\description{
Calculate Geyer (1992) NSE estimator.
}
\details{
The type "iseq" is a wrapper around \link[mcmc]{initseq} from the 
MCMC package and gives the positive intial sequence estimator.
 The type "bm" is the batch mean estimator.
 The type "obm" is the overlapping batch mean estimator.
 The type "iseq.bm" is a combinaison of the two.
}
\examples{
n    = 1000
ar   = 0.9
mean = 1
sd   = 1
set.seed(1234)
x = as.vector(arima.sim(n = n, list(ar = ar), sd = sd) + mean)
nse.geyer(x = x, type = "bm", nbatch = 30)
nse.geyer(x = x, type = "obm", nbatch = 30)
nse.geyer(x = x, type = "iseq", iseq.type = "pos")
nse.geyer(x = x, type = "iseq.bm", iseq.type = "con")
 
}
\references{
Geyer, Charles J. "Practical markov chain monte carlo." Statistical Science (1992): 473-483.
}

