% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R
\name{nse.spec0}
\alias{nse.spec0}
\title{The spectral density at zero.}
\usage{
nse.spec0(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The variance estimator.
}
\description{
Calculate the variance of the mean with the spectrum at zero estimator.
}
\details{
This is a wrapper around \link[coda]{spectrum0.ar} form the CODA package.
}
\examples{
n = 1000
ar = c(0.9)
mean = c(1)
sd = c(10)

Ts1 = as.vector(arima.sim(n = n, list(ar = ar), sd = sd) + mean)

nse::nse.spec0(x = Ts1)
}
\references{
Plummer, Martyn, et al. "CODA: Convergence diagnosis and output analysis for MCMC." R news 6.1 (2006): 7-11.
}

