% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{nse.spec0}
\alias{nse.spec0}
\title{The spectral density at zero.}
\usage{
nse.spec0(x, method = c("AR", "bartlett", "wosa", "tukey"),
  prewhite = FALSE, max.length = 200)
}
\arguments{
\item{x}{A numeric vector.}

\item{method}{A character string denoting the method to use in estimating the spectral density function}

\item{prewhite}{Prewhite the serie before analysis}

\item{max.length}{maximum sample size for aggregation}
}
\value{
The variance estimator.
}
\description{
Calculate the variance of the mean with the spectrum at zero estimator.
}
\details{
This is a wrapper around \link[coda]{spectrum0.ar} from the CODA package, \link[sapa]{SDF} from the sapa package, \link[psd]{psdcore} from the PSD package and \link[mcmcse]{mcse}from the mcmcse package.
}
\examples{
n = 1000
ar = c(0.9)
mean = c(1)
sd = c(10)
 
Ts1 = as.vector(arima.sim(n = n, list(ar = ar), sd = sd) + mean)
 
nse::nse.spec0(x = Ts1, method = "AR", prewhite = FALSE)
 
}
\references{
Plummer, Martyn, et al. "CODA: Convergence diagnosis and output analysis for MCMC." R news 6.1 (2006): 7-11.

D.B. Percival and A. Walden "Spectral Analysis for Physical Applications: Multitaper and Conventional Univariate Techniques". Cambridge University Press (1993).

Barbour, A. J. and R. L. Parker (2014), "psd: Adaptive, sine multitaper power spectral density estimation for R", Computers & Geosciences Volume 63 February 2014 : 1-8

James M. Flegal, John Hughes and Dootika Vats. (17-08-2015). mcmcse: Monte Carlo Standard Errors for MCMC. R package version 1.1-2. Riverside, CA and Minneapolis, MN.
}

