% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{nse.spec0}
\alias{nse.spec0}
\title{The spectral density at zero.}
\usage{
nse.spec0(x, method = c("AR", "lag window", "wosa", "multitaper"))
}
\arguments{
\item{x}{A numeric vector.}

\item{method}{A character string denoting the method to use in estimating the spectral density function}
}
\value{
The variance estimator.
}
\description{
Calculate the variance of the mean with the spectrum at zero estimator.
}
\details{
This is a wrapper around \link[coda]{spectrum0.ar} from the CODA package and \link[sapa]{SDF} from the sapa package.
}
\examples{
n = 1000
ar = c(0.9)
mean = c(1)
sd = c(10)
 
Ts1 = as.vector(arima.sim(n = n, list(ar = ar), sd = sd) + mean)
 
nse::nse.spec0(x = Ts1, method = "AR")
 
}
\references{
Plummer, Martyn, et al. "CODA: Convergence diagnosis and output analysis for MCMC." R news 6.1 (2006): 7-11.

D.B. Percival and A. Walden "Spectral Analysis for Physical Applications: Multitaper and Conventional Univariate Techniques". Cambridge University Press (1993).
}

